# src/public/swagger/swagger.yaml
openapi: 3.0.3
info:
  title: Personal Blog
  description: Personal Blog
  termsOfService: http://swagger.io/terms/
  contact:
    email: bangvo.5sense.vn@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0
servers:
  - url: http://localhost:8555/api
    description: Local server
...
components:
  schemas:
...
    ApiResponse:
      type: object
      properties:
        status:
          type: integer
          description: HTTP status code
        message:
          type: string
          description: Human-readable message
        data:
          type: object
          description: Response data, can be specific to each endpoint
        notFound:
          type: object
          description: Information for 404 (Not Found) response
          properties:
            resource: 
              type: string
              description: Name or type of the resource that was not found
        internalServerError:
          type: object
          description: Information for 500 (Internal Server Error) response
          properties:
            error:
              type: string
              description: Details of the internal server error
        invalidInput:
          type: object
          description: Information for 400 (Bad Request - Invalid Input) response
          properties:
            fields:
              type: array
              items:
                type: string
                description: List of fields with invalid input values
        unauthorized:
          type: object
          description: Information for 401 (Unauthorized) response
          properties:
            error:
              type: string
              description: Details of the unauthorized request
...
    CreatePublisherRequest:
      type: object
      properties:
        name:
          type: string
          example: "Publisher X"
        address:
          type: string
          example: "123 Main Street, City"
      required:
        - name
        - address
...
  securitySchemes:
    token:
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
...
  /publishers/create:
    post:
      tags:
        - Publisher Management
      summary: Create a new publisher
      operationId: createPublisher
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePublisherRequest'
            example:
              name: "Publisher X"
              address: "123 Main Street, City"
      responses:
        '201':
          description: Publisher created successfully
          content:
            application/json:
              example:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Bad request - Invalid input
          content:
            application/json:
              example:
                $ref: '#/components/schemas/ApiResponse'
      security:
        - token: []
  /publishers:
    get:
      tags:
        - Publisher Management
      summary: Get all publishers
      operationId: getAllPublishers
      responses:
        '200':
          description: List of publishers
          content:
            application/json:
              example:
                $ref: '#/components/schemas/ApiResponse'
        '404':
          description: Publishers not found
          content:
            application/json:
              example:
                $ref: '#/components/schemas/ApiResponse'
      security:
        - token: []
...